To be continued

+++ src/map.rs
@@ -74,8 +74,8 @@ impl Map {
 
         for i in i1..=i2 {
             for j in j1..=j2 {
-                let x = (i * self.block_size) as i32 - x1 as i32;
-                let y = (j * self.block_size) as i32 - y1 as i32;
+                let x = (i * self.block_size) as f32 - x1 as f32;
+                let y = (j * self.block_size) as f32 - y1 as f32;
                 let b = self.block(i, j);
                 self.block_tiles[b as usize].add(([x as f32, y as f32],));
             }
